Adds utility to export countylookup results to CSV (#3)
* Adds utility to export countylookup results to CSV

Implemented exportToCSV function that takes the zipToCounty map and writes the data to a generated CSV file, properly formatted with clear headers.

Tested:
 - Ran the program with a valid tar file and observed output in the generated CSV.
 - Checked error messages for scenarios with a missing tar file argument.
 - Checked error messages for scenarios with an invalid tar file.

* Refactors countylookup to write results to stdout.

This refactor enhances flexibility by writing results to stdout instead of a CSV file. Users can now redirect output as needed. Additionally, the user can specify a single ZIP code to print its associated counties directly.

Tested:
Ran the program with and without the -zip flag and observed the output.
Checked error messages for scenarios with a missing -tar flag.
Checked error messages for scenarios with an invalid tar file.

* Folded logic from `printCountiesForZip`

Collapsed the logic of `printCountiesForZIP` into main. Ensured the program exits with a non-zero status code if the user-specified zip is not found.

Tested:
    Ran the program with and without the -zip flag and observed the output.
    Checked error messages for scenarios with a missing -tar flag.
    Checked error messages for scenarios with an invalid tar file.

---------

Co-authored-by: Tamara Dement <tamaradement@Tamaras-Laptop.local>